Fix implicit down cast in var number reading.

+++ src/network/types/VarNumber.hh
@@ -24,7 +24,7 @@ public:
       if (it == end)
         return {false,end,0};
       read = *(it++);
-      result |=  (read & 0x7F) << (7 * indx++);
+      result |=  static_cast<std::int32_t>(read & 0x7F) << (7 * indx++);
       if (indx > 5) throw std::overflow_error("VarInt is too big");
     }
     while (read & 0x80);
@@ -46,7 +46,7 @@ public:
       if (it == end)
         return {false,begin,0};
       read = *(it++);
-      result |=  (read & 0x7F) << (indx++ * 7);
+      result |=  static_cast<std::int64_t>(read & 0x7F) << (indx++ * 7);
       if (indx > 10) throw std::overflow_error("VarLong is too big");
     }
     while (read & 0x80);
